Updated Usermodule and Assistance module

+++ BookingSystem/Repository/IBookandPaymentRepository.cs
@@ -17,8 +17,8 @@ namespace BookingSystem.Repository
 
         Task UpdateBookingAsync(long BookingID, DateTime StartDate);
 
-        Task DeleteBookingAsync(int BookingID);
-        Task CancelBooking(int BookingID);
+        Task DeleteBookingAsync(long BookingID);
+        Task CancelBooking(long BookingID);
